move router to baes Layout
